api_config_prod/requirements.txt
altair==4.0.0
argh==0.26.2
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==1.0.3
bleach==3.1.0
blinker==1.4
boto3==1.11.0
botocore==1.14.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
idna==2.8
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.5
jupyterlab-server==1.0.6
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.1.0
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
numpy==1.18.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
s3transfer==0.3.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
streamlit==0.52.2
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.7
validators==0.14.1
watchdog==0.9.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
xgboost==0.90
zipp==0.6.0
